+++ examples/25-c99/helloworld.c
@@ -43,7 +43,7 @@ int _main_(int _argc, char** _argv)
 	while (!entry_process_events(&width, &height, &debug, &reset) )
 	{
 		// Set view 0 default viewport.
-		bgfx_set_view_rect(0, 0, 0, width, height);
+		bgfx_set_view_rect(0, 0, 0, (uint16_t)width, (uint16_t)height);
 
 		// This dummy draw call is here to make sure that view 0 is cleared
 		// if no other draw calls are submitted to view 0.
@@ -51,13 +51,14 @@ int _main_(int _argc, char** _argv)
 
 		// Use debug font to print information about this example.
 		bgfx_dbg_text_clear(0, false);
-		bgfx_dbg_text_image(uint16_max(width/2/8, 20)-20
-						 , uint16_max(height/2/16, 6)-6
-						 , 40
-						 , 12
-						 , s_logo
-						 , 160
-						 );
+		bgfx_dbg_text_image(
+			  uint16_max( (uint16_t)width /2/8, 20)-20
+			, uint16_max( (uint16_t)height/2/16, 6)-6
+			, 40
+			, 12
+			, s_logo
+			, 160
+			);
 		bgfx_dbg_text_printf(0, 1, 0x4f, "bgfx/examples/25-c99");
 		bgfx_dbg_text_printf(0, 2, 0x6f, "Description: Initialization and debug text with C99 API.");
